implemented masking of flagged data.
Data that is flagger.BAD is masked by default.
One can pass a flag or a list-like to the param to_mask to control the behavior.
Pass an empty list to disable masking.
Nevertheless IFF the index changed the masked data is not restored.
Also flags are NOT restored IFF they considered _flagged_ before the func call, but not afterwards anymore.
Also any data that is set on masked locations is not overwritten by unmasking.