fixed logging, added debug message indicates the test that run.
i removed the setup of the logging module from the core. It interferes
with the needs of the user of the lib (saqc). eg. Messages are double
printed or not at all.

>> It is strongly advised that you do not add any handlers other than
>> NullHandler to your library’s loggers. [...]
from https://docs.python.org/3/howto/logging.html

we might add a defalutLoggingSetup(log_level='INFO') function or similar in future,
to ease the setup the logging.